pyproject.toml
[project]
name = "life-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.88.0",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "pocketflow>=0.0.3",
    "python-dotenv>=1.2.2",
]
